[Compositor] Support for splitted refsw and nxclient include directories
For Yocto refsw headers are installed in /usr/include/refsw and /usr/include/nxclient. However the current cmake expects (a Buildroot generate install) all headers to be in /usr/include/refsw.
This will result into the following error when building in RDKV/Yocto:
    NexusServer.cpp:418:106: error: 'nxserverlib_init_extended' was not declared in this scope
    |                      _instance = nxserverlib_init_extended(&_serverSettings, config.Authentication.Value());

Updated the cmake to look for both refsw and nxclient subdirs and set the LIBNEXUS_INCLUDE_DIRS accordingly.

Signed-off-by: wouterlucas <wouter@wouterlucas.com>