[CMAKE][RemoteControl][RF4CE] Provide an option to disable RF4CE functionality. By default turn on to maintain previous behaviour.
In some cases (e.g. faulty GP unit on a dev-box) it is quite annoying that RF4CE gets enable by default if the GP bits are found. With the previous functionality I'd have to forcefully remove the RF4CE bits from the staging area in order to get the build to work. Instead, lets do a flag and make it a little easier to turn RF4CE on or off.

Signed-off-by: wouterlucas <wouter@wouterlucas.com>