[IRRemote] Always read 2 items out of the queue, to avoid misalignment if we miss the first callback.
This fixes an issue with an BCM7250 17.3 broadcom box which 1 out of 5 bootups missed the first callback. By missing the first call back and only asking for 1 item the queue the queue was always mis-aligned and we end up reading the repeat key. That resulted into duplicate keys (as the repeat key gets read first) on the UI or key missed (if you change keys). By reading 2 items from the queue, if we ever misalign, we automatically realign by reading just enough to empty the queue in this particular case.

Signed-off-by: wouterlucas <wouter@wouterlucas.com>