Compositor: unregistering standby acknowledgment
when the device is set to standby, nexus server waits for all its
connected clients to acknowledge the standby request. The device will
go to standby only on receiving an acknowledge or after reaching a fixed
timeout. This commit will unregister plugins using compositor from the client list
from which nexus server is expecting the acknowledgment.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>