[Compositor] Make NxServer Audio Loudness Mode configurable
The Audio Loudness Mode being set when starting the Nexus Server is
currently hardcoded to 'NEXUS_AudioLoudnessEquivalenceMode_eNone'.
Different usage scenarios may required another configuration. For
example, there's currently a requirement to use a different mode
('NEXUS_AudioLoudnessEquivalenceMode_eEbuR128') for passing a Dolby
certification.

The best approach is to let this parameter be a plugin configuration
(e.g. like the HDCP Mode), so make the changes to allow this. For
keeping backwards compatibility, the default mode is still kept as
'NEXUS_AudioLoudnessEquivalenceMode_eNone'.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>