[Streamer] Fix ExternalAccess implementation for latest Thunder
The ExternalAccess class, extending RPC::Communicator class, is used by
plugins to allow external RPC access to their services.

With latest Thunder came some changes in the way the ExternalAccess
class must be implemented, as can be seen, e.g., by tracing the changes
made to the implementation done by the OpenCDMi and Compositor plugins.

These adjustments are missing in the Streamer plugin, which causes a
SIGSEGV once a client opens an RPC channel to it. Fix it by applying the
same adjustments that were made to the ExternalAccess implementation for
the OpenCDMi and Compositor plugins.

This suggests that it might be worth having an ExternalAccess class
template defined in Thunder's core, to prevent having multiple copies of
almost identical ExternalAccess class implementations spread all over
the plugins' code.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>