[Streamer] Register to RPC::IRemoteConnection::INotification
The plugins should register to RPC::IRemoteConnection::INotification in
order to ensure that a plugin gets successfully deactivated in case its
remote connection dies.

This is what ensures that, e.g., a plugin gets successfully deactivated
if it crashes. Without this, the core will still "see" the plugin as
being active after it has crashed.

To be noted that the Deactivated() implementation in the Streamer
already existed, but was not being used at all, since it was not
registering to get these notifications.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>