Update: RemoteControl: Add changes to support sendign events
This Commit add teh implementation of a new interface IRemoteControl
to the RemoteControl plugin.

The new interface is meant to be used to register with the plugin for
any events from any of the KeyProducers defined in RemoteControl plugin
like GreenPeak("RF4CE"), IRRemote..etc.

Also add the changes to get the Pairing related events from GreenPeak
RF4CE implementation to registered clients.

Events are received by IKeyHandler from KeyProducers, just as we are
doing for Keys, And this event is passed to IRemoteControl interface who
will finally pass the event to all the registered clients.

Signed-off-by: harikrishnan.p <harikris2012@gmail.com>