[RemoteControl] Fix registering of device specific map file
Commit 4002913b272b ("[BLUETOOTHCONTROL] Got it compiling again, with
ADPCM and PCM decoders.") removed use of 'TCHAR *' in the RemoteControl
core. However, in doing that, it broke the string that must be used as
key for registering the device's specific map file in the input handler.

For example, for the RF4CE device, instead of registering the specific
map file with the "RF4CE" key it started to register the specific map
file using a key that is the map file name. With this, the input handler
will then fail to locate a map file for the producers upon key events,
because the producer name doesn't match with any table, because the
table was registered using the map file name instead of the producer
name.

This happened because the commit removed the 'producer' variable and now
just uses the 'loadName' variable. The 'producer' contained the producer
name (e.g. "RF4CE"), was never changed during the method execution and
was later as the key to register the specific map file. The 'loadName'
starts by having the producer name, but then changes to have the map
file name instead. So, when registering with 'loadName', it ends up
registering with the map file name instead of the producer name.

Fix this by restoring the use of the 'producer' variable for keeping the
producer name.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>