RemoteControl: Make PassOn field configurable for default producer
Change the PassOn field in RemoteControl.json configurable so that we
can select whether the KeyCode needs to be mapped or not build-time for
the default producer.

Previously the passon field was not configured in the
RemoteControl.json file.

The passon field can be set to true to use "default" key producer to
simulate key presses.

Signed-off-by: harikrishnan.p <harikris2012@gmail.com>