RemoteControl: Fix: Dont set Pair twice
This commit updates the Pairing logic for RF4CE producer, A new check is
added to ensure that we dont try to set the device to pairing mode
twice.

i.e if the device is already in pairing mode, Dont allow the Pair() call
to set the device to pairing mode again.

It was observed that if we set the device in pairing mode to pairing
mode again, The pairing will not succed. Thus this check is neccessary
to ensure that we don't try to set a device in pairing mode to pairing
mode again.

Signed-off-by: harikrishnan.p <harikris2012@gmail.com>