RemoteControl: GreenPeak: Add Logic for Autopairing
This commit adds the needed logic for AutoPairing RF4CE RCU in GreenPeak
module.

We are using gpRf4ce_GetNrOfPairingEntries to get the number of paired
RCU to find whether we need to initiate the AutoPairing sequence.

While the RemoteControl plugin inits we will check whether there are any
paired RCU units, If there are no paired RCU units we will initiate the
AutoPairing Sequence

The check is also done when we get pairing related events. This is to
ensure that after we get pairing failed or pairing timedout event we
will keep the device in pairing mode again.

The AutoPairing logic is flagged using a config variable from the
RemoteControl.json.

By default the AutoPairing feature is turned off.

Signed-off-by: harikrishnan.p <harikris2012@gmail.com>