[IOConnector] Fix handler's Start/Stop integer size
All internal variables for dealing with the handler's 'Start' and 'Stop'
settings are 'uint32_t', but the associated members in the respective
config class are 'Core::JSON::DecUInt8', which results in undesired
values (wrap around, truncated) for values larger than 255 (MAX_UINT8).

This way, use 'Core::JSON::DecUInt32' instead on the config class.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>