[IOConnector] Increase handler's Start/Stop integer size
All internal variables for dealing with the handler's 'Start' and 'Stop'
settings are 'uint32_t' for representing the time in milliseconds.  The
associated members in the respective config class are
'Core::JSON::DecUInt8' for seconds, allowing for a maximum of 255
(MAX_UINT8) seconds and undesired values (wrap around, truncated) for
values larger than 255.

This way, use 'Core::JSON::DecUInt16' instead on the config class,
allowing larger values. Allows for a maximum of 65535 (MAX_UINT16)
seconds while still not overflowing the 'uint32_t' representation in
milliseconds (65535000 = 0x3e7fc18).

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>