[BLUETOOTHREMOTE] By default, use the set keymap.
Currently there are two remotes around that have the same keymap. The remotes are from
Tech4Home and are called Notes S1 and Notes S2. If you do not specify a keymap file, the
name is used of the remote to find an associated keymap. The current deployment was only
with a Notes S2.json file and the keymap was not set.
This commit makes sure that:
1) The configuratin keymap file is used.
2) Create a default keymapfile called "bluetooth.json" that is always
   used since it is default now set in the config.

This means that the Notes S1 and Notes S2 remotes should work.

I have been testing all the button and saw that on the remote I have the keys [play/pause]
and [home] where sending differnt key codes than the initialy issued remotes. I changed
the codes accordingly to work.

The remote I am using announces itself as Notes S1 and has firmware version: 0197.00.03