[Compositor] Nexus: Allow configuring list of real clients
Add a new JSON config entry named "allowedclients", for Nexus compositor
only, which can be used to tell the compositor which are the names of
the clients that it should accept. When this entry is set to a non-empty
list, the Nexus compositor only accepts a client if its name is in this
list, otherwise it behaves as nameless client, i.e., it accepts the
connection but doesn't process it in terms of compositing.

By default, this entry will not exist to keep backwards compatibility.
It can be set using the 'PLUGIN_COMPOSITOR_ALLOWED_CLIENTS' cmake
variable.

Not all Nexus clients (SW modules to do a Nexus Join) are real
compositor clients, so they should not be accounted for compositing
purposes, like z-order management, input focus, etc. The Nexus
compositor is considering any joined client as a compositing client as
long as it joins with some name. In scenarios where clients are
joining/un-joining, this is causing problems where the compositor ends
up giving focus (z-order, etc.) to Nexus clients that are not managing
any video/graphics surfaces, thus taking away the focus from real
clients that need it.

One of the problematic scenarios happens when EspialBrowser (a real
compositing client) is playing YouTube content and then some other
(non-compositing) clients are registering/unregistering. In this case,
Z-Order management ends up giving focus to a non-compositing client
which ends up incorrectly hiding the video layer from the playing
YouTube. This problem doesn't happen if the non-composting clients are
not registered in the compositor.

The new option allows telling exactly what are the clients that are real
compositing ones. It could also be possible to have the list telling
which ones are not to register, but it's easier to say which ones are,
as there is no control over all the Nexus clients out there.

This change doesn't break previous behavior. By default, it is kept.
Only when the "allowedclients" entry is specified to a non-empty list
will the compositor now validate the client names against it.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>