RDKDEV-257: Fix config for PLUGIN_POWER_GPIOTYPE and PLUGIN_POWER_GPIOPIN
Make sure that the cmake if statements for PLUGIN_POWER_GPIOTYPE and
PLUGIN_POWER_GPIOPIN are only applicable to the configuration key/values
they control, and not the whole configuration.

There can be cases where the Power plugin needs to be configured with
the GPIO feature disabled (meaning those variables won't be set) while
still requiring the other configuration values given by, e.g.,
PLUGIN_POWER_POWER_KEY and PLUGIN_POWER_CONTROL_CLIENTS, so the cmake if
statements for PLUGIN_POWER_GPIOTYPE and PLUGIN_POWER_GPIOPIN should
only be used for the keys they control, and not the whole configuration.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>