RDKDEV-257: Fix disabling of the ControlClients feature
The ControlClients config setting was introduced so that it is possible
to disable the plugin feature consisting of maintaining a list of
plugins implementing IStateControl (plugins that can be suspended /
resumed) and then directly suspending those plugins when going to
standby and resuming them when leaving standby.

With the feature disabled, the plugins are no longer suspended /
resumed, which is the intended behavior. However, the Power plugin is
still doing a lot of stuff that doesn't need to and should not do with
the feature disabled. This includes listening to plugin state changes
and then querying them for IStateControl, register with itself for
getting power state notifications, etc. None of these are required and
are actually causing problems with plugins deactivation, even with the
feature disabled.

This way, make sure that the feature is completely disabled when
configure that way:

 * Do not register to receive nor process plugin state notifications.
 * Do not register to receive nor process power state notifications.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>