[COMPOSITOR] Cleanup.
The compositor plugin is split-upo in two parts. One is the generic part that takes care of the
communication over JSONRPC/COMRPC to the actual compositor (code in Compositor/ directory) and
the actual compositor. This is code found in Compositor/lib/<implementation>. The first part
communication should have *NO* compositor logic, that is the responsibility of the Compositor
implementation.
This cleanup moves the Z-Order logic out of the Compositor communication part into the RPI
compositor implementation. The Wayland (Westeros/Weston implementations) already have this
logic. The Nexus implementation has this functionality still missing. Needs to be added in
the future.