[INPUTSWITCH] Enable a COM interface and JSONRPC to "Select" a channel.
In stead of iterating over all the output channels from the Key Handler in Thunder and treating
the status of each channel independently, allow for a method called Select (IInputSwitch/JSONRPC).
This method enables the indicated channel and disables all the other (non-immune) channels.

A channel can be made immune by explicitly specifing ENABLE/DISABLE for such a channel. The
immunity can be removed by setting it back to SLAVE. This automatically disables the channels.
Immunity is also removed from a channel if it is used in the Select call.