[DRMPlayer] Implemented decryption for video
This change is meant to provide a baseline implementation
for a gstreamer client using OpenCDM for decryption. The current
limitations do not allow for playback of multliple streams from
the same source.

The current implementation allows for:
  - Playback of a single encrypted stream (video or audio).
  - Both widevine and playready streams are accepted.

As in most gstreamer clients, generating dotfiles is happening
on each state change.