[WIFICONTROLLER] The state behaviour of the Connect was not correct.
After receiving an OK from the last request, it still had to wait for a CONNECT event or an ERROR report. To
include this in the state machine of the connect (implementation details that needs to be hidden in the
controller) added a WAIT state, during this WaitState the CONNECT Event, or the SSID temporarly unavailable
is expected and should move the ConnectRequest to completion.
This is needed for the AutoConnect functionality in the WifiController and for regular Connects, issued through
the JSONRPC/Restfull interface.