WebServer: Decouple the proxy from the plugin's availability
In the original implementation, the proxy for DIALServer and Device info
is enabled by default when these plugins are enabled. There are cases
where we need proxy to be enabled to any one of the component
irrespective of the other plugins state. For example, in a system where
DIAL and DeviceInfo is enabled, and dial is configured to use the proxy server,
Webserver need to expose the port to the external world. In this case,
DeviceInfo is also get exposed to the external world. In order to avoid
this, the config has been tweaked to use a different config variable to
enable the proxy.