WebServer: Makes file server configurable
This commit is for making the file serving feature configurable.
This is needed since, in cases we only wants the WebServer to be a
proxy for another plugin, we don't want to exposes the filesystem
contents to external world. In such case, we could use this to disable
the file serving feature.
With this change, we would be able to configure WebServer only as a
proxy server.