[CLEANUP] Followup of the Module/rework cleanup on Thunder.
As there is no need to use plugins as librraies by other plugins, there is no need
to define any class to be externally accessible. This commit cleans all EXTERNAL
keywords still available on classes. This should be a very low risk as all the
Module.h files would already undefine EXTERNAL and set it to an empty value :-)