Added AMLogic implementation
This commit introduces a reworked version of the DisplayInfo
implementation for an AMLogic hp40a platform. The notifications
about a state change of the connection are handled via
incoming udev messages.

TODO:
Currently, the udev state change messages are not being generated
on a succesfull authentication. The lack of udev messages
is being investigated. Tracking link:
https://jira.rdkcentral.com/jira/browse/HP40A-603