IOConnector: Fix pin observer list notified when marker is reached
Without this change, when a marker is reached, the ExternalBase::Updated() function is invoked
and the clients registered in the ExternalBase through the GPIO::Subscribe(Exchange::IExternal::INotification*)
function are notified. This notification is incorrect as the clients being notified are not the ones registered
by the IInputPin::Register method.
This change updates the observer list that is notified.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>