IOConnector: Update GPIO event handling
With the current implementation, the check for GPIO state changes is executed by pooling (see WPEFramework::GPIO::Pin::Handle() method),
and additionally, it can also be performed periodically (minimum periodicity of 1 sec) with the use of a WPEFramework::Exchange::ExternalBase::Timed instance.
The first one, serves to notify the list of handlers (registered to the GPIO through the ExternalBase) that a GPIO state change occured, the periodic checking
(see WPEFramwork::GPIO::Pin::Evaluate()) verifies if a GPIO state change occurred, and if so, it notifies both the ExternalBase handlers and the observers
registered through the IInputPin abstract base class.

For an observer to register himself to a specific Pin instance and add markers to be notified when reached, the IOConnector plugin must enable periodic GPIO state check
 (will be performed concurrently with the pooling check).  The issue of having both pooling and periodic checks is the verification of state changes:
bool Pin::HasChanged() const
{
        return (Get() != _lastValue);
}

If the Handle() (pooling check) is called before Evaluate() (the periodic check), it will notify its clients that a change occurred and the Pin _lastValue data member
is aligned to the current value _lastValue = Get();,  this assignment  interferes with Evaluate()  expected behavior as HasChanged() will return false and consequently
the markers management and observers notification will not occur.

To ensure that the markers management and observers notification are performed, this commit presents a workaround where the periodic check is dropped and the Evaluate()
invocation is performed within Handle().

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>