IOConnector: Fix and rework to enable the use of the TimedPin  (#523)
* IOConnector: Add capability to obtain a specific GPIO::Pin base class pointer

This commit overrides the IInputPin resource method to allow services to query for a specific GPIO::Pin instance through RPC.
This commit also resolves ambiguity issues between ICatalog base classes.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>

* IOConnector: Fix GPIO::Pin IInputPin::Unregister override

This commit serves to fix the implementation of the GPIO::Pin IInputPin::Unregister override.
The fix changes the TimedPin method called, by calling TimedPin::Unregister instead of
TimedPin::Register. With this, the sink handle reference counting is decreased and removed from the
list of observers (expected behavior).

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>

* IOConnector: Fix pin observer list notified when marker is reached

Without this change, when a marker is reached, the ExternalBase::Updated() function is invoked
and the clients registered in the ExternalBase through the GPIO::Subscribe(Exchange::IExternal::INotification*)
function are notified. This notification is incorrect as the clients being notified are not the ones registered
by the IInputPin::Register method.
This change updates the observer list that is notified.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>

* IOConnector: Update behavior when duplicated markers are to be added

This commit updates the TimedInput behavior when the Add function is called
with a duplicated marker. Instead of terminating the process on debug builds
and inserting the duplicate marker non-debug builds, a condition is added to
only insert non-duplicated markers into the TimedInput markers list.

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>

* IOConnector: Update GPIO event handling

With the current implementation, the check for GPIO state changes is executed by pooling (see WPEFramework::GPIO::Pin::Handle() method),
and additionally, it can also be performed periodically (minimum periodicity of 1 sec) with the use of a WPEFramework::Exchange::ExternalBase::Timed instance.
The first one, serves to notify the list of handlers (registered to the GPIO through the ExternalBase) that a GPIO state change occured, the periodic checking
(see WPEFramwork::GPIO::Pin::Evaluate()) verifies if a GPIO state change occurred, and if so, it notifies both the ExternalBase handlers and the observers
registered through the IInputPin abstract base class.

For an observer to register himself to a specific Pin instance and add markers to be notified when reached, the IOConnector plugin must enable periodic GPIO state check
 (will be performed concurrently with the pooling check).  The issue of having both pooling and periodic checks is the verification of state changes:
bool Pin::HasChanged() const
{
        return (Get() != _lastValue);
}

If the Handle() (pooling check) is called before Evaluate() (the periodic check), it will notify its clients that a change occurred and the Pin _lastValue data member
is aligned to the current value _lastValue = Get();,  this assignment  interferes with Evaluate()  expected behavior as HasChanged() will return false and consequently
the markers management and observers notification will not occur.

To ensure that the markers management and observers notification are performed, this commit presents a workaround where the periodic check is dropped and the Evaluate()
invocation is performed within Handle().

Signed-off-by: Joel Pinto <joelpostman97@gmail.com>