IOCOnnector: Fix issue where notification is not sent for last marker in list
It is observed that for the last marker in '_markers' list,
the observers are not notified when that particular marker is reached.
This is due to the wrong check added in Reached().
This commit fix the issue by correcting the logic.

This commit also brings in the changes to return 'false' from
'Reached' when '_markers' list is empty as there is no need to
notify observers in this case.