Bug fix and feature enhancement for Power plugin (#580)
* RDKDEV-257: power: Update power state change notification to include previous state

This commit is for changing the interface of the power state change
notification. This is needed to include the information about the
previous state state in the power state change notification

SetState() in Power.cpp has the logic to send power state
change notification (before changing power state).
But framework resources may be blocked during this time period and it may delay
other calls that can happen at the same time from other threads.

This commit removes the logic in SetState() in Power.cpp to send power state change notification.
It is now implementation's choice/reposnsibility to send "Before" power state change notification.

Implementation is using Workerpool job to change power state.
Sending "Before" power state change notification from this Job will also
ensure that framework resources are not blocked during this time.

* RDKDEV-257: Fix config for PLUGIN_POWER_GPIOTYPE and PLUGIN_POWER_GPIOPIN

Make sure that the cmake if statements for PLUGIN_POWER_GPIOTYPE and
PLUGIN_POWER_GPIOPIN are only applicable to the configuration key/values
they control, and not the whole configuration.

There can be cases where the Power plugin needs to be configured with
the GPIO feature disabled (meaning those variables won't be set) while
still requiring the other configuration values given by, e.g.,
PLUGIN_POWER_POWER_KEY and PLUGIN_POWER_CONTROL_CLIENTS, so the cmake if
statements for PLUGIN_POWER_GPIOTYPE and PLUGIN_POWER_GPIOPIN should
only be used for the keys they control, and not the whole configuration.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>

* RDKDEV-257: Fix disabling of the ControlClients feature

The ControlClients config setting was introduced so that it is possible
to disable the plugin feature consisting of maintaining a list of
plugins implementing IStateControl (plugins that can be suspended /
resumed) and then directly suspending those plugins when going to
standby and resuming them when leaving standby.

With the feature disabled, the plugins are no longer suspended /
resumed, which is the intended behavior. However, the Power plugin is
still doing a lot of stuff that doesn't need to and should not do with
the feature disabled. This includes listening to plugin state changes
and then querying them for IStateControl, etc.
None of these are required and are actually causing problems
with plugins deactivation, even with the feature disabled.

This way, to make sure that the feature is completely disabled,
configure to neither register to receive nor process plugin state notifications.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>

* RDKDEV-257: Power: Correct key for controlclients

Key name for controlclients is added as "controlclients" in Power.config.
But it is added as "control" in Power::Config::Config().
This mismatch causes failure in reading the value of controlclients from Power.json.

This commit fix the issue by correcting the key.

Co-authored-by: Johan Saji <johansaji@tataelxsi.co.in>
Co-authored-by: Ricardo Silva <ricardo.silva@gbtembedded.com>