WebServer: Add support to work with the SecurityAgent plugin (#660)
In case the SecurityAgent plugin is enabled and activated, this
WebServer plugin needs to obtain and use a WebToken to authenticate
itself to Thunder when acting as a proxy to other Thunder services, like
the DIALServer plugin. Without this, DIAL will fail when the
SecurityAgent is enabled.

This way, add a new option ('PLUGIN_WEBSERVER_SECURITYAGENT') that when
enabled, makes the WebServer plugin obtain a 'localhost' WebToken and
use it when proxying the configured requests, so that these proxied
requests can be authenticated and succeed.

The new option requires a dependency to the 'securityagent' Thunder
Client Library and a compile time definition ('ENABLE_SECURITY_AGENT')
to determine whether it needs to call the library's GetToken() function
to get the token or not. When enabled, the new option will also set a
new runtime configuration 'securityagent' key to 'true', so that the
plugin knows that it needs to call GetToken(). As a result, the plugin
also needs the new 'Security' precondition, so that it's guaranteed that
the token can be obtained when the WebServer plugin is activated.

Finally, when the support for the SecurityAgent is enabled, the
WebServer will add a new 'Authorization' header to the web requests
before servicing them to the final destination (e.g. DIALServer),
containing the token obtained with GetToken().

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>