IOConnector: Handle key release without key press scenario
When an interrupt is triggered for a pin, the current value is fetched
to determine if it was a key press or release.

It is observed that during bootup, the interrupt gets triggered and since
there was no actual key press made, the pin value would be '0' and this
would be treated as a key release.
So in this case there was no previous key press corresponding to this
key release.

This may cause unexpected behaviors like unnecessarily triggering the
handle for the markers.

To fix this issue, before processing key release event, ensure that
 '_pressedTime' is non-zero to ensure that there was previous key press event.
After processing key release event, reset '_pressedTime' to 0.