[NetworkControl] MAC control dynamically.. (#764)
* [NETWORKCONTROL] Allow the MAC address to be condigured and set, if needed.

* [MAC] MAC control dynamically..

The BPI (Banana Pi BPI-M2 ZERO), has a wired and Wifi stack. The MAC address assigned to these
NIC's is however, by default randomly generated and thus changes on each boot.
The MAC could be fixed in several ways:
1) Set a MAC in the DTS overlay (how to make sure it is unique ??)
2) Set a MAC through UBOOT (jpw tp make sure it is unique ??)
3) Enhance the NetworkControl plugin :-)

This commmit does just that. The NetworkControl plugin can now have a "mac" field per interface. The "mac" field
can hold a mac address like "AA:BB:CC:DD:EE:FF" or it can be set to null.

If it contains a "real" MAC address, it will set that MAC address on the interface before starting it (if it is
not already started). To resolve the unique issue, you can also set null for the mac field, than the NetworkControl
plugin will check if it had run before (there is in the PersistentStorage space a file with the last run information)
and use the MAC address from that time (effectively mainting the first auto generated MAC address.

Problem Solved !

* Update DHCPClient.h

* Make sure MACSize is declared in source too

---------

Co-authored-by: HaseenaSainul <41037131+HaseenaSainul@users.noreply.github.com>
Co-authored-by: HaseenaSainul <h.sainul@metrological.com>