[COMPOSITOR] Do not assert on configuration errors, continue looking for connectors.
Avoid that the compositor asserts if the configuration is incorrect, just report it can not operate during initialization, than using loggin or tracing
the root casue should be to detect that the configuration is incorrect, fix it and go on.

On the Bananapi, the first connector (CTC) found did not have a buffer, now continue to search if you fain another one with a buffer. No clue if this is recommended :-)