[NETWORKCONTROL] Also cope with DNS files that have a size of 0..
If the /etc/resolve.conf file (typeically the DNS file) does not exist or is empty, in debug, an
ASSERT fires in the DateElement since one triees to access an empty buffer.
While fixing this also found that there is already a Search for pattern method in the DataElement, so
start using it. Prune where you can prune ;-)