Development/compositor lifetime fixes (#880)
* [COMPOSITOR] Have the External/Remote usage of the ICompoition::IClient a forwarded class.

As the IComposition::IClient interface is used by the CompositorClient to access a surface, the lifetime of this usage,
usage in a different process, needs to be unlinked from the references maintained in the compositor implementation. As
the CompositorClient usage should be leading in the lifetime of the surface/client implemented and used in the compositor,
the Release() of the IComposition::IClient in the CompositorClint libraries now triggers a dedicated Release (of a Remote
Client) in the compositor Implementation. Getting this signal (if it is the "last" Remote Client, now triggers the "real"
implementation of the client to actually Revoke and cleanup from the compositor (Revoke()).

* Update Compositor.cpp

* [Compositor] Corrected member initialization.

* [COMPOSITOR] Fix error return code

* [COMPOSITOR] Fix reference counting logic and move texture release to Revoke method

---------

Co-authored-by: Pierre Wielders <pierre@wielders.net>