[COMPOSITOR] IRendering extend with possibility to pass the Destination texture.
Since some buffers (Output buffers, DMA) can have two buffers to quickly swap, rendering to the output
is not always fixed to one buffer.
Initially the IRendering was created with the one buffer. This has been removed and now for each rendering
the Texture where to render to must be passed. Allowing the multi buffer output (DRM) to determine on which
buffer should be used for rendering wihout constantly creating and deleting the buffer.